pera-sam/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.12
librosa==0.10.2
numpy==1.26.4
# tensorflow-cpu: CPU-only build (~200MB vs 620MB for full tensorflow)
# Sufficient for autoencoder inference — no GPU required
tensorflow-cpu==2.17.1
PyYAML==6.0.2
scikit-learn==1.5.2
tqdm==4.66.5
matplotlib==3.9.2
requests==2.32.3
